add generate code hash
