+++ api/src/utils/types.ts
@@ -1,0 +1,3 @@
+export function transformTypes(types: object): any {
+  return Object.values(types).reduce((res, types): object => ({ ...res, ...types }), {});
+}
